// src/ui/SelectRole/SelectRole.interface.ts
import { ControllerRenderProps } from 'react-hook-form';
import { Options } from 'react-select';
import { IFieldProps } from '../Field/Field.type';

export interface IOption {
  label: string;
  value: string;
}

export interface ISelect extends IFieldProps {
  options: Options<IOption>;
  isMulti?: boolean;
  field: ControllerRenderProps<any, any>;
  isLoading?: boolean;
}